requirements.txt
dash==2.14.1
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
PyQt6==6.6.1
python-dotenv==1.0.0
TA-Lib==0.4.28
schwabdev>=2.5.0 